In test IO utility, restore the old stdin/stdout
Instead of restoring `sys.stdin` to `sys.__stdin__`, restore it to
whatever it was before we installed out dummy I/O hooks. This is
relevant in pytest, for example, which installs its *own* `sys.stdin`,
which we were then clobbering. This was leading to the suppression of
test failures observed in #5021 and addressed in #5027.